[#266] - scrutinizer corrections and minor refactoring
