+++ src/Html/Link/AbstractLink.php
@@ -16,10 +16,10 @@ namespace Phalcon\Html\Link;
 use Phalcon\Support\Collection;
 
 /**
- * @property array  $attributes
- * @property string $href
- * @property array  $rels
- * @property bool   $templated
+ * @property Collection $attributes
+ * @property string     $href
+ * @property Collection $rels
+ * @property bool       $templated
  */
 abstract class AbstractLink
 {
