+++ src/Image/Adapter/AbstractAdapter.php
@@ -331,26 +331,7 @@ abstract class AbstractAdapter implements AdapterInterface
         int $height = null,
         int $master = Enum::AUTO
     ): AdapterInterface {
-        switch ($master) {
-            case Enum::TENSILE:
-            case Enum::AUTO:
-            case Enum::INVERSE:
-            case Enum::PRECISE:
-                if (null === $width || null === $height) {
-                    throw new Exception("width and height must be specified");
-                }
-                break;
-            case Enum::WIDTH:
-                if (null === $width) {
-                    throw new Exception("width must be specified");
-                }
-                break;
-            case Enum::HEIGHT:
-                if (null === $height) {
-                    throw new Exception("height must be specified");
-                }
-                break;
-        }
+        $this->checkResizeInput($width, $height, $master);
 
         if ($master !== Enum::TENSILE) {
             if ($master === Enum::AUTO) {
@@ -387,8 +368,10 @@ abstract class AbstractAdapter implements AdapterInterface
             }
         }
 
-        $width  = (int) max(round($width), 1);
-        $height = (int) max(round($height), 1);
+        $width   = null === $width ? 0 : $width;
+        $height  = null === $height ? 0 : $height;
+        $width   = (int) max(round($width), 1);
+        $height  = (int) max(round($height), 1);
 
         $this->processResize($width, $height);
 
@@ -543,150 +526,6 @@ abstract class AbstractAdapter implements AdapterInterface
     }
 
     /**
-     * @param int $red
-     * @param int $green
-     * @param int $blue
-     * @param int $opacity
-     *
-     * @return void
-     */
-    abstract protected function processBackground(
-        int $red,
-        int $green,
-        int $blue,
-        int $opacity
-    ): void;
-
-    /**
-     * @param int $radius
-     *
-     * @return void
-     */
-    abstract protected function processBlur(int $radius): void;
-
-    /**
-     * @param int $width
-     * @param int $height
-     * @param int $offsetX
-     * @param int $offsetY
-     *
-     * @return void
-     */
-    abstract protected function processCrop(
-        int $width,
-        int $height,
-        int $offsetX,
-        int $offsetY
-    ): void;
-
-    /**
-     * @param int $direction
-     *
-     * @return void
-     */
-    abstract protected function processFlip(int $direction): void;
-
-    /**
-     * @param AdapterInterface $mask
-     *
-     * @return void
-     */
-    abstract protected function processMask(AdapterInterface $mask);
-
-    /**
-     * @param int $amount
-     *
-     * @return void
-     */
-    abstract protected function processPixelate(int $amount): void;
-
-    /**
-     * @param int  $height
-     * @param int  $opacity
-     * @param bool $fadeIn
-     *
-     * @return void
-     */
-    abstract protected function processReflection(
-        int $height,
-        int $opacity,
-        bool $fadeIn
-    ): void;
-
-    /**
-     * @param string $extension
-     * @param int    $quality
-     *
-     * @return false|string
-     * @throws Exception
-     */
-    abstract protected function processRender(string $extension, int $quality);
-
-    /**
-     * @param int $width
-     * @param int $height
-     *
-     * @return void
-     */
-    abstract protected function processResize(int $width, int $height): void;
-
-    /**
-     * @param int $degrees
-     *
-     * @return void
-     */
-    abstract protected function processRotate(int $degrees): void;
-
-    /**
-     * @param string $file
-     * @param int    $quality
-     *
-     * @return void
-     * @throws Exception
-     */
-    abstract protected function processSave(string $file, int $quality): void;
-
-    /**
-     * @param int $amount
-     *
-     * @return void
-     */
-    abstract protected function processSharpen(int $amount): void;
-
-    /**
-     * @param string      $text
-     * @param mixed       $offsetX
-     * @param mixed       $offsetY
-     * @param int         $opacity
-     * @param int         $red
-     * @param int         $green
-     * @param int         $blue
-     * @param int         $size
-     * @param string|null $fontFile
-     *
-     * @return void
-     * @throws Exception
-     */
-    abstract protected function processText(
-        string $text,
-        mixed $offsetX,
-        mixed $offsetY,
-        int $opacity,
-        int $red,
-        int $green,
-        int $blue,
-        int $size,
-        string $fontFile = null
-    ): void;
-
-    abstract protected function processWatermark(
-        AdapterInterface $watermark,
-        int $offsetX,
-        int $offsetY,
-        int $opacity
-    ): void;
-
-    /**
      * @param int $value
      * @param int $min
      * @param int $max
@@ -696,5 +535,42 @@ abstract class AbstractAdapter implements AdapterInterface
     protected function checkHighLow(int $value, int $min = 0, int $max = 100): int
     {
         return min($max, max($value, $min));
+    }
+
+    /**
+     * Resize the image to the given size
+     *
+     * @param int|null $width
+     * @param int|null $height
+     * @param int      $master
+     *
+     * @return void
+     * @throws Exception
+     */
+    private function checkResizeInput(
+        int $width = null,
+        int $height = null,
+        int $master = Enum::AUTO
+    ): void {
+        switch ($master) {
+            case Enum::TENSILE:
+            case Enum::AUTO:
+            case Enum::INVERSE:
+            case Enum::PRECISE:
+                if (null === $width || null === $height) {
+                    throw new Exception("width and height must be specified");
+                }
+                break;
+            case Enum::WIDTH:
+                if (null === $width) {
+                    throw new Exception("width must be specified");
+                }
+                break;
+            case Enum::HEIGHT:
+                if (null === $height) {
+                    throw new Exception("height must be specified");
+                }
+                break;
+        }
     }
 }
