+++ src/Image/Adapter/Gd.php
@@ -605,7 +605,7 @@ class Gd extends AbstractAdapter
                 );
         }
 
-        return ob_get_clean();
+        return (string) ob_get_clean();
     }
 
     /**
